+++ src/zanthropic.zig
@@ -3,6 +3,7 @@ const provider = @import("zigai").providers.anthropic;
 pub const api_base = provider.api_base;
 pub const api_version = provider.api_version;
 pub const Error = provider.Error;
+pub const Provider = provider.Provider;
 pub const Client = provider.Client;
 pub const encodeRequest = provider.encodeRequest;
 pub const encodeStreamingRequest = provider.encodeStreamingRequest;