+++ src/zgoogle.zig
@@ -2,6 +2,7 @@ const provider = @import("zigai").providers.google;
 
 pub const api_base = provider.api_base;
 pub const Error = provider.Error;
+pub const Provider = provider.Provider;
 pub const Client = provider.Client;
 pub const encodeRequest = provider.encodeRequest;
 pub const decodeResponse = provider.decodeResponse;